Add files via upload
This is the [water_potability dataset](https://www.kaggle.com/datasets/adityakadiwal/water-potability?resource=download) from Kaggle. The file contains water quality metrics for 3276 different water bodies. It appears to be the same data we originally discussed, though I'm not sure if it's the original Kaggle posting we reviewed. It mentions that the data is synthetically generated.